Enum amplify_syn::ArgValue

pub enum ArgValue {
    Literal(Lit),
    Type(Type),
    Expr(Expr),
    None,
}

Value for attribute or attribute argument, i.e. for #[attr = value] and #[attr(arg = value)] this is the value part of the attribute. Can be either a single literal or a single valid rust type name

Variants

Literal(Lit)

Attribute value represented by a literal

Type(Type)

Attribute value represented by a type name

Expr(Expr)

Attribute value represented by an expression

None

No value is given

Implementations

impl ArgValue

pub fn into_literal_value(self) -> Result<Lit, Error>

Converts into literal value for ArgValue::Literal variant or fails with Error::ArgValueMustBeLiteral otherwise

pub fn into_type_value(self) -> Result<Type, Error>

Converts into type value for ArgValue::Type variant or fails with Error::ArgValueMustBeType otherwise

pub fn into_type_expr(self) -> Result<Expr, Error>

Converts into type value for ArgValue::Expr variant or fails with Error::ArgValueMustBeExpr otherwise

pub fn to_literal_value(&self) -> Result<Lit, Error>

Constructs literal value for ArgValue::Literal variant or fails with Error::ArgValueMustBeLiteral otherwise

pub fn to_type_value(&self) -> Result<Type, Error>

Constructs type value for ArgValue::Type variant or fails with Error::ArgValueMustBeType otherwise

pub fn to_type_expr(&self) -> Result<Expr, Error>

Constructs type value for ArgValue::Expr variant or fails with Error::ArgValueMustBeExpr otherwise

pub fn as_literal_value(&self) -> Option<&Lit>

Returns a reference to a literal value for ArgValue::Literal variant, or None otherwise.

pub fn as_type_value(&self) -> Option<&Type>

Returns a reference to a literal value for ArgValue::Type variant, or None otherwise.

pub fn as_type_expr(&self) -> Option<&Expr>

Returns a reference to a literal value for ArgValue::Expr variant, or None otherwise.

pub fn is_none(&self) -> bool

Tests whether the self is set to ArgValue::None

pub fn is_some(&self) -> bool

Tests whether the self is not set to ArgValue::None

pub fn value_class(&self) -> Option<ValueClass>

Returns ValueClass for the current value, if any