Enum amplify_syn::ArgValueReq

pub enum ArgValueReq {
    Required {
        default: Option<ArgValue>,
        class: ValueClass,
    },
    Optional(ValueClass),
    Prohibited,
}

Requirements for attribute or named argument value presence

Variants

Required

Fields

default: Option<ArgValue>

Default value

class: ValueClass

Type of the value literal

Argument must hold a value with the provided class

Optional(ValueClass)

Argument or an attribute may or may not hold a value

Prohibited

Argument or an attribute must not hold a value

Implementations

impl ArgValueReq

pub fn with_default(default: impl Into<ArgValue>) -> ArgValueReq

Constructs argument requirements object with default value

pub fn required(class: ValueClass) -> ArgValueReq

Construct ArgValueReq::Required variant with no default value

pub fn value_class(&self) -> Option<ValueClass>

Returns value class requirements, if any

pub fn default_value(&self) -> ArgValue

Returns default argument value. If not default is provided within the requirement, returns ArgValue::None (since this is de facto default value for any argument).

pub fn is_required(&self) -> bool

Determines whether argument is required to have a value

pub fn check(
&self,
value: &mut ArgValue,
attr: impl ToString,
arg: impl ToString
) -> Result<(), Error>

Checks the argument against current requirements, generating Error if the requirements are not met.