Enum amplify_syn::TypeClass

pub enum TypeClass {

    Array,
    BareFn,
    Group,
    ImplTrait,
    Infer,
    Macro,
    Never,
    Paren,
    Path,
    Ptr,
    Reference,
    Slice,
    TraitObject,
    Tuple,
    Verbatim,

}

Constrains for the possible types that a Rust value could have.

Variants

Array

A fixed size array type: [T; n].

BareFn

A bare function type: fn(usize) -> bool.

Group

A type contained within invisible delimiters.

ImplTrait

An impl Bound1 + Bound2 + Bound3 type where Bound is a trait or a lifetime.

Infer

Indication that a type should be inferred by the compiler: _.

Macro

A macro in the type position.

Never

The never type: !.

Paren

A parenthesized type equivalent to the inner type.

Path

A path like std::slice::Iter, optionally qualified with a self-type as in <Vec<T> as SomeTrait>::Associated.

Ptr

A raw pointer type: *const T or *mut T.

Reference

A reference type: &'a T or &'a mut T.

Slice

A dynamically sized slice type: [T].

TraitObject

A trait object type Bound1 + Bound2 + Bound3 where Bound is a trait or a lifetime.

Tuple

A tuple type: (A, B, C, String).

Verbatim

Tokens in type position not interpreted by Syn.

Implementations

impl TypeClass

pub fn check(
self,
ty: &Type,
attr: impl ToString,
arg: impl ToString
) -> Result<(), Error>

Checks the Type against current requirements, generating Error if the requirements are not met.