Enum amplify_syn::ArgValue

pub enum ArgValue {
    Literal(Lit),
    Type(Type),
    None,
}

Value for attribute or attribute argument, i.e. for #[attr = value] and #[attr(arg = value)] this is the value part of the attribute. Can be either a single literal or a single valid rust type name

Variants

Literal(Lit)

Attribute value represented by a literal

Type(Type)

Attribute value represented by a type name

None

No value is given

Implementations

impl ArgValue

pub fn literal_value(&self) -> Result<Lit, Error>

Returns literal value for ArgValue::Literal variant or fails with Error::ArgValueMustBeLiteral otherwise

pub fn type_value(&self) -> Result<Type, Error>

Returns type value for ArgValue::Type variant or fails with Error::ArgValueMustBeType otherwise

pub fn is_none(&self) -> bool

Tests whether the self is set to ArgValue::None

pub fn is_some(&self) -> bool

Tests whether the self is not set to ArgValue::None

pub fn value_class(&self) -> Option<ValueClass>

Returns ValueClass for the current value, if any